Version 6
Updated nodes and launch files to set nav goals according to the specific waypoints within the map. Was able to pass basic course north and south with this code. Still need to tune the costmap parameters to pass advanced courses. Also, the locations are set to the center of the waypoints right now. We can always adjust this if it's advantageous to do so. I also just picked what I thought was a logical order for the advanced course but we can always change that order, as well.